update: App.jsx / index.jsx 파일 구분
