+++ src/App.jsx
@@ -1,0 +1,28 @@
+import React, { useState } from 'react';
+
+import uuid from 'react-uuid';
+
+import Input from './Input';
+import Todo from './Todo';
+
+export default function App() {
+  const [todoItem, setTodoItem] = useState([]);
+
+  const handleAddTodo = (item) => {
+    const result = [...todoItem];
+    result.push(item);
+    setTodoItem(result);
+  };
+
+  return (
+    <div>
+      <p>To-do</p>
+      <Input handleClick={handleAddTodo} />
+      <ul>
+        {todoItem.map((item) => (
+          <Todo lable={item} key={uuid()} id={uuid()} />
+        ))}
+      </ul>
+    </div>
+  );
+}
